backend/requirements.txt
# Django core
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0,<25.0

# Database
psycopg[binary]>=3.1,<4.0
dj-database-url>=2.1,<3.0

# Cache & Sessions
django-redis>=5.4,<6.0

# Task Queue
celery>=5.3,<6.0
redis>=5.0,<6.0

# Auth & Security
djangorestframework-simplejwt>=5.3,<6.0
django-allauth>=0.61,<1.0
cryptography>=42.0,<44.0

# Payments
stripe==14.3.0

# Email
resend>=2.23.0,<3.0
svix>=1.21,<2.0
requests>=2.31,<3.0

# AI Providers
anthropic>=0.40.0,<1.0
openai>=1.50,<2.0

# Storage (S3-compatible for MinIO)
django-storages>=1.14,<2.0
boto3>=1.34,<2.0

# PDF generation
weasyprint>=62.0,<63.0
pydyf>=0.10.0,<0.11.0

# Image handling
Pillow>=10.2,<11.0

# WSGI server
gunicorn>=21.0,<22.0

# Environment
python-dotenv>=1.0,<2.0

# Dev tools
django-extensions>=3.2,<4.0
django-debug-toolbar>=4.2,<5.0

# RUM Pipeline (Kafka, ClickHouse, OpenTelemetry)
kafka-python-ng>=2.0.2,<3.0
clickhouse-connect>=0.7.0,<1.0
opentelemetry-api>=1.23.0,<2.0
opentelemetry-sdk>=1.23.0,<2.0
opentelemetry-exporter-otlp>=1.23.0,<2.0
opentelemetry-instrumentation-django>=0.44b0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.7,<5.0
factory-boy>=3.3,<4.0
